requirements.txt
pandas==2.1.0
sqlalchemy==2.0.23
sentence-transformers==2.2.2
faker==19.3.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
pytest==7.4.2
pyyaml==6.0.1
joblib==1.3.2
torch==2.0.1
redis==5.0.1
azure-storage-blob==12.18.3
azure-identity==1.15.0
azure-cosmos==4.5.1
opencensus-ext-azure==1.1.10
chromadb==0.4.15 